refactor(DBConnection): Connection 분리
