+++ src/main/java/dao/DBConnection.java
@@ -1,0 +1,23 @@
+package dao;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DBConnection {
+
+    private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
+    private static final String DATABASE = "chess"; // MySQL DATABASE 이름
+    private static final String OPTION = "?useSSL=false&serverTimezone=UTC";
+    private static final String URL = "jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION;
+    private static final String USERNAME = "root"; //  MySQL 서버 아이디
+    private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
+
+    public Connection getConnection() {
+        try {
+            return DriverManager.getConnection(URL, USERNAME, PASSWORD);
+        } catch (final SQLException e) {
+            throw new IllegalStateException("데이터베이스 연결에 실패했습니다!");
+        }
+    }
+}
